+++ lab4/shares.h
@@ -71,8 +71,8 @@ extern TaskShare<uint16_t>* sh_encoder_error_count_1;		// Motor 1
 extern TaskShare<uint16_t>* sh_encoder_error_count_2;		// Motor 2
 
 // Motor PID setpoints
-extern TaskShare<int16_t>* sh_setpoint_1;			// Motor 1
-extern TaskShare<int16_t>* sh_setpoint_2;			// Motor 2
+extern TaskShare<int32_t>* sh_setpoint_1;			// Motor 1
+extern TaskShare<int32_t>* sh_setpoint_2;			// Motor 2
 
 // Motor Power values from PID control
 extern TaskShare<int16_t>* sh_PID_1_power;			// Motor 1
@@ -80,5 +80,6 @@ extern TaskShare<int16_t>* sh_PID_2_power;			// Motor 2
 
 // Flag to indicate PID control enabled
 extern TaskShare<uint8_t>* sh_PID_control;
+
 
 #endif /// _SHARES_H_
